+++ src/NServiceBus.Core/Transports/Development/DirectoryBasedTransaction.cs
@@ -5,18 +5,6 @@ namespace NServiceBus
     using System.IO;
     using System.Linq;
 
-    interface IDevelopmentTransportTransaction
-    {
-        string FileToProcess { get; }
-
-        void BeginTransaction(string incomingFilePath);
-        void Commit();
-        void Rollback();
-        void ClearPendingOutgoingOperations();
-        void Enlist(string messagePath, List<string> messageContents);
-        void Complete();
-    }
-
     class DirectoryBasedTransaction : IDevelopmentTransportTransaction
     {
         public DirectoryBasedTransaction(string basePath)
